+++ CS386Project/app/build.gradle
@@ -4,8 +4,8 @@ android {
     compileSdkVersion 29
     buildToolsVersion "29.0.3"
     defaultConfig {
-        applicationId "com.example.cs386project"
-        minSdkVersion 22
+        applicationId "com.example.csapp"
+        minSdkVersion 21
         targetSdkVersion 29
         versionCode 1
         versionName "1.0"
