+++ CS386Project/app/src/androidTest/java/com/example/csapp/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.cs386project;
+package com.example.csapp;
 
 import android.content.Context;
 
@@ -22,6 +22,6 @@ public class ExampleInstrumentedTest {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
 
-        assertEquals("com.example.cs386project", appContext.getPackageName());
+        assertEquals("com.example.csapp", appContext.getPackageName());
     }
 }
