+++ CS386Project/app/src/main/res/values/strings.xml
@@ -1,6 +1,7 @@
 <resources>
-    <string name="app_name">CS386Project</string>
-    <string name="action_settings">Settings</string>
-    <string name="the_project_the_game">The Project: the game</string>
-    <string name="startgame">Start Game</string>
+    <string name="app_name">CSApp</string>
+    <string name="csproject_the_game">CSProject: The Game</string>
+    <string name="start">Start</string>
+    <string name="title_activity_new_game_page">NewGamePage</string>
+    <string name ="charactertext"> Welcome create your character. </string>
 </resources>
