+++ CS386Project/app/src/test/java/com/example/csapp/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.cs386project;
+package com.example.csapp;
 
 import org.junit.Test;
