+++ CS386Project/settings.gradle
@@ -1,2 +1,2 @@
 include ':app'
-rootProject.name='CS386Project'
+rootProject.name='CSApp'
